* In case of a 'for' loop, don't declare the iterator of a type that has a larger value range when the iterator can take a maximum value of a more efficient data type. (int to uint8 in most cases). * Other minor cleanups, Null pointer checks etc.
--HG--
branch : trunk

+++ src/bindings/scripts/ScriptMgr.cpp
@@ -894,7 +894,7 @@ void ScriptsFree()
     delete []SpellSummary;
 
     // Free resources before library unload
-    for(int i=0;i<MAX_SCRIPTS;i++)
+    for(uint16 i =0;i<MAX_SCRIPTS;i++)
         delete m_scripts[i];
 
     num_sc_scripts = 0;
@@ -934,7 +934,7 @@ void ScriptsInit(char const* cfg_file = "trinitycore.conf")
     bar.step();
     outstring_log("");
 
-    for(int i=0;i<MAX_SCRIPTS;i++)
+    for(uint16 i =0;i<MAX_SCRIPTS;i++)
         m_scripts[i]=NULL;
 
     FillSpellSummary();
