+++ src/bindings/scripts/scripts/zone/karazhan/boss_moroes.cpp
@@ -62,7 +62,7 @@ struct TRINITY_DLL_DECL boss_moroesAI : public ScriptedAI
 {
     boss_moroesAI(Creature *c) : ScriptedAI(c)
     {
-        for(int i = 0; i < 4; i++)
+        for(uint8 i = 0; i < 4; i++)
         {
             AddId[i] = 0;
         }
@@ -185,7 +185,7 @@ struct TRINITY_DLL_DECL boss_moroesAI : public ScriptedAI
             }
         }else
         {
-            for(int i = 0; i < 4; i++)
+            for(uint8 i = 0; i < 4; i++)
             {
                 Creature *pCreature = m_creature->SummonCreature(AddId[i], Locations[i][0], Locations[i][1], POS_Z, Locations[i][2], TEMPSUMMON_CORPSE_TIMED_DESPAWN, 10000);                
                 if (pCreature)
@@ -198,7 +198,7 @@ struct TRINITY_DLL_DECL boss_moroesAI : public ScriptedAI
 
     bool isAddlistEmpty()
     {
-        for(int i = 0; i < 4; i++)
+        for(uint8 i = 0; i < 4; i++)
         {
             if(AddId[i] == 0)
                 return true;
